+++ src/commands.rs
@@ -1,0 +1,36 @@
+use std::time::Duration;
+
+use core_bluetooth::central::AdvertisementData;
+use core_bluetooth::central::characteristic::Characteristic;
+use core_bluetooth::central::peripheral::Peripheral;
+use core_bluetooth::central::service::Service;
+use core_bluetooth::error::Error;
+use core_bluetooth::uuid::Uuid;
+
+#[derive(Debug, Clone)]
+pub enum CommandResult {
+    GetStatus(Duration, Option<usize>),
+    ListPeripherals(Vec<(Peripheral, AdvertisementData)>),
+    FindPeripheral(Option<(Peripheral, AdvertisementData)>),
+    ConnectToPeripheral(Peripheral),
+    FindService(Peripheral, Option<Service>),
+    FindCharacteristic(Peripheral, Option<Characteristic>),
+    ReadCharacteristic(Peripheral, Characteristic, Option<Vec<u8>>),
+    WriteCharacteristic(Peripheral, Characteristic, Result<(), Error>),
+}
+
+#[derive(Debug, Clone)]
+pub enum Command {
+    GetStatus,
+    ListPeripherals,
+    FindPeripheralByService(Uuid),
+    ConnectToPeripheral((Peripheral, AdvertisementData)),
+    RegisterConnectedPeripheral(Peripheral),
+    UnregisterConnectedPeripheral(Uuid),
+    FindService(Peripheral, String),
+    FindCharacteristic(Peripheral, Service, String, uuid::Uuid),
+    // last uuid is not BLE uuid, but internal uuid used for hashmap
+    ReadCharacteristic(Peripheral, Characteristic),
+    WriteCharacteristic(Peripheral, Characteristic, Vec<u8>),
+}
+
