+++ src/handlers.rs
@@ -1,14 +1,11 @@
 use std::collections::HashMap;
-use std::fmt::Debug;
 use std::process::exit;
 use std::time::{Duration, Instant};
 
 use async_std::{task, task::JoinHandle};
 use core_bluetooth::{central::{*,
-                               characteristic::{Characteristic, WriteKind},
-                               peripheral::Peripheral,
-                               service::Service},
-                     error::Error,
+                               characteristic::WriteKind,
+                               peripheral::Peripheral},
                      ManagerState,
                      uuid::Uuid};
 use futures::{select, StreamExt};
@@ -16,242 +13,8 @@ use log::*;
 use postage::{*, prelude::Sink};
 use postage::mpsc::Sender;
 
-// --- Implementation draft ----------------------------------------------------
-// https://stackoverflow.com/questions/27957103/how-do-i-create-a-heterogeneous-collection-of-objects
-// https://play.rust-lang.org/?version=stable&mode=debug&edition=2018&gist=308ca372ab13fdb3ecec6f4a3702895d
-
-
-#[derive(Debug, Clone)]
-pub enum CommandResult {
-    GetStatus(Duration, Option<usize>),
-    ListPeripherals(Vec<(Peripheral, AdvertisementData)>),
-    FindPeripheral(Option<(Peripheral, AdvertisementData)>),
-    ConnectToPeripheral(Peripheral),
-    FindService(Peripheral, Option<Service>),
-    FindCharacteristic(Peripheral, Option<Characteristic>),
-    ReadCharacteristic(Peripheral, Characteristic, Option<Vec<u8>>),
-    WriteCharacteristic(Peripheral, Characteristic, Result<(), Error>),
-}
-
-#[derive(Debug, Clone)]
-pub enum Command {
-    GetStatus,
-    ListPeripherals,
-    FindPeripheralByService(Uuid),
-    ConnectToPeripheral((Peripheral, AdvertisementData)),
-    RegisterConnectedPeripheral(Peripheral),
-    UnregisterConnectedPeripheral(Uuid),
-    FindService(Peripheral, String),
-    FindCharacteristic(Peripheral, Service, String, uuid::Uuid),
-    // last uuid is not BLE uuid, but internal uuid used for hashmap
-    ReadCharacteristic(Peripheral, Characteristic),
-    WriteCharacteristic(Peripheral, Characteristic, Vec<u8>),
-}
-// OPTIMIZE(df): Move matchers to separate file?
-
-enum EventMatch {
-    Next(Command),
-    Result(CommandResult),
-    None,
-}
-
-impl EventMatch {
-    fn is_none(&self) -> bool {
-        matches!(*self, EventMatch::None)
-    }
-}
-
-// it is critical EventMatchers are stateless
-
-trait EventMatcher {
-    fn matches(&self, event: &CentralEvent) -> EventMatch;
-}
-
-struct PeripheralDiscoveredMatcherByServiceUUID {
-    uuid: Uuid,
-}
-
-impl PeripheralDiscoveredMatcherByServiceUUID {
-    fn new(uuid: Uuid) -> Self {
-        Self { uuid }
-    }
-}
-
-impl EventMatcher for PeripheralDiscoveredMatcherByServiceUUID {
-    fn matches(&self, event: &CentralEvent) -> EventMatch {
-        if let CentralEvent::PeripheralDiscovered { peripheral, advertisement_data, rssi: _ } = event {
-            if advertisement_data.service_uuids().iter()
-                .find(|&uuid| uuid.eq(&self.uuid)).is_some() {
-                return EventMatch::Result(CommandResult::FindPeripheral(Some((peripheral.clone(), advertisement_data.clone()))))
-            }
-        }
-        EventMatch::None
-    }
-}
-
-struct PeripheralConnectedMatcher {
-    peripheral_id: Uuid,
-}
-
-impl PeripheralConnectedMatcher {
-    fn new(peripheral: &Peripheral) -> Self {
-        Self {
-            peripheral_id: peripheral.id()
-        }
-    }
-}
-
-impl EventMatcher for PeripheralConnectedMatcher {
-    fn matches(&self, event: &CentralEvent) -> EventMatch {
-        if let CentralEvent::PeripheralConnected { peripheral } = event {
-            if peripheral.id() == self.peripheral_id {
-                return EventMatch::Next(Command::RegisterConnectedPeripheral(peripheral.clone()))
-            }
-        } else if let CentralEvent::PeripheralConnectFailed { peripheral, error } = event {
-            if peripheral.id() == self.peripheral_id {
-                let id = peripheral.id().clone();
-                error!("failed to connect to peripheral {:?}: {:?}", peripheral, error);
-                return EventMatch::Next(Command::UnregisterConnectedPeripheral(id))
-            }
-        }
-        EventMatch::None
-    }
-}
-
-
-struct ServicesDiscoveredMatcher {
-    peripheral_id: Uuid,
-    uuid_substr: String,
-}
-
-impl ServicesDiscoveredMatcher {
-    fn new(peripheral: &Peripheral, uuid_substr: String) -> Self {
-        Self {
-            peripheral_id: peripheral.id(),
-            uuid_substr,
-        }
-    }
-}
-
-impl EventMatcher for ServicesDiscoveredMatcher {
-    fn matches(&self, event: &CentralEvent) -> EventMatch {
-        if let CentralEvent::ServicesDiscovered { peripheral, services } = event {
-            if peripheral.id() == self.peripheral_id {
-                return EventMatch::Result(CommandResult::FindService(
-                    peripheral.clone(),
-                    find_by_id_substr(&self.uuid_substr, services)))
-            }
-        }
-        EventMatch::None
-    }
-}
-
-
-struct CharacteristicsDiscoveredMatcher {
-    peripheral_id: Uuid,
-    service_id: Uuid,
-    uuid_substr: String,
-}
-
-impl CharacteristicsDiscoveredMatcher {
-    fn new(peripheral: &Peripheral, service: &Service, uuid_substr: String) -> Self {
-        Self {
-            peripheral_id: peripheral.id(),
-            service_id: service.id(),
-            uuid_substr,
-        }
-    }
-}
-
-trait HasId {
-    fn get_id(&self) -> Uuid;
-}
-
-impl HasId for Service {
-    fn get_id(&self) -> Uuid {
-        self.id().clone()
-    }
-}
-
-impl HasId for Characteristic {
-    fn get_id(&self) -> Uuid {
-        self.id().clone()
-    }
-}
-
-fn find_by_id_substr<T>(id_substr: &String, items: &Result<Vec<T>, Error>) -> Option<T> where T: HasId + Clone {
-    let id_substr = id_substr.as_str();
-    return items.as_ref().ok()
-        .and_then(|items| items.iter().find(|item| item.get_id().to_string().contains(id_substr)))
-        .map(|item| item.clone());
-}
-
-impl EventMatcher for CharacteristicsDiscoveredMatcher {
-    fn matches(&self, event: &CentralEvent) -> EventMatch {
-        if let CentralEvent::CharacteristicsDiscovered { peripheral, service, characteristics } = event {
-            if peripheral.id() == self.peripheral_id && service.id() == self.service_id {
-                return EventMatch::Result(CommandResult::FindCharacteristic(
-                    peripheral.clone(),
-                    find_by_id_substr(&self.uuid_substr, characteristics)))
-            }
-        }
-        EventMatch::None
-    }
-}
-
-struct CharacteristicValueMatcher {
-    peripheral_id: Uuid,
-    characteristic_id: Uuid,
-}
-
-impl CharacteristicValueMatcher {
-    fn new(peripheral: &Peripheral, characteristic: &Characteristic) -> Self {
-        Self {
-            peripheral_id: peripheral.id(),
-            characteristic_id: characteristic.id(),
-        }
-    }
-}
-
-impl EventMatcher for CharacteristicValueMatcher {
-    fn matches(&self, event: &CentralEvent) -> EventMatch {
-        if let CentralEvent::CharacteristicValue { peripheral, characteristic, value } = event {
-            if peripheral.id() == self.peripheral_id && characteristic.id() == self.characteristic_id {
-                return EventMatch::Result(CommandResult::ReadCharacteristic(
-                    peripheral.clone(),
-                    characteristic.clone(),
-                    if let Ok(value) = value { Some(value.clone()) } else { None }))
-            }
-        }
-        EventMatch::None
-    }
-}
-
-struct WriteCharacteristicResultMatcher {
-    peripheral_id: Uuid,
-    characteristic_id: Uuid,
-}
-
-impl WriteCharacteristicResultMatcher {
-    fn new(peripheral: &Peripheral, characteristic: &Characteristic) -> Self {
-        Self {
-            peripheral_id: peripheral.id(),
-            characteristic_id: characteristic.id(),
-        }
-    }
-}
-
-impl EventMatcher for WriteCharacteristicResultMatcher {
-    fn matches(&self, event: &CentralEvent) -> EventMatch {
-        if let CentralEvent::WriteCharacteristicResult { peripheral, characteristic, result } = event {
-            if peripheral.id() == self.peripheral_id && characteristic.id() == self.characteristic_id {
-                return EventMatch::Result(
-                    CommandResult::WriteCharacteristic(peripheral.clone(), characteristic.clone(), result.clone()));
-            }
-        }
-        EventMatch::None
-    }
-}
+use crate::commands::*;
+use crate::event_matchers::*;
 
 pub struct Controller {
     sender: mpsc::Sender<(Command, oneshot::Sender<CommandResult>)>,
