+++ src/lib.rs
@@ -2,3 +2,5 @@ pub mod handlers;
 pub mod shutting_down_stream;
 pub mod input_token;
 pub mod byte_slices;
+pub mod commands;
+mod event_matchers;
